updated id

+++ pkg/table/table64/table64.go
@@ -13,7 +13,7 @@ import (
 func New(start, end uint64) table.Table {
 	return &table64{
 		table: idxtable.NewTable[tree.Entry](
-			int64(end - start + 1),
+			uint64(end - start + 1),
 		),
 		start: start,
 		end:   end,
@@ -140,7 +140,7 @@ func (r *table64) GetByLabel(selector labels.Selector) tree.Entries {
 		entry := iter.Value().Data()
 		if selector.Matches(entry.Labels()) {
 			// need to remap the id for the outside world
-			entry := tree.NewEntry(id64.NewID(uint64(calculateIDFromIndex(r.start, int64(entry.ID().ID()))), id64.IDBitSize), entry.Labels())
+			entry := tree.NewEntry(id64.NewID(uint64(calculateIDFromIndex(r.start, uint64(entry.ID().ID()))), id64.IDBitSize), entry.Labels())
 			entries = append(entries, entry)
 		}
 	}
@@ -157,11 +157,11 @@ func (r *table64) validateID(id uint64) error {
 	return nil
 }
 
-func calculateIndex(id, start uint64) int64 {
+func calculateIndex(id, start uint64) uint64 {
 	// Calculate the index in the bitmap
-	return int64(id - start)
+	return uint64(id - start)
 }
 
-func calculateIDFromIndex(start uint64, id int64) uint64 {
-	return start + uint64(id)
+func calculateIDFromIndex(start uint64, id uint64) uint64 {
+	return start + id
 }
